backend/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
Django==5.2.6
django-cors-headers==4.7.0
django-rest-knox==5.0.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dynaconf==3.2.11
hvac==2.3.0
idna==3.10
knox==0.1.14
loguru==0.7.3
psycopg2-binary==2.9.10
PyJWT==2.10.1
PyMySQL==1.1.2
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
shortuuid==1.0.13
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
validators==0.35.0
win32_setctime==1.2.0
cryptography==46.0.2
bcrypt===5.0.0
